requirements.txt
fastapi
uvicorn
python-telegram-bot
httpx
pydantic
pyyaml
python-dotenv
